+++ models/DynamicModel.php
@@ -7,7 +7,7 @@ use wdmg\helpers\ArrayHelper;
  * Yii2 DynamicModel
  *
  * @category        Model
- * @version         1.2.0
+ * @version         1.2.1
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-base
  * @copyright       Copyright (c) 2019 - 2020 W.D.M.Group, Ukraine