+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+## 1.2.1 (2020-05-02)
+ * Added SluggableBehavior(), implement for base ActiveRecord
+
 ## 1.2.0 (2020-04-14)
  * Added getModule() for base ActiveRecord model
  * Removing DynamicModel to models